Refactoring Rebus in Portuguese (Brazil) and Reset Button v4
